requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.0.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
keyboard==0.13.5
phonenumbers==8.12.36
Pillow==8.4.0
psycopg2-binary==2.9.1
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
